+++ lib/survey-gizmo-ruby.rb
@@ -11,6 +11,7 @@ require 'digest/md5'
 require 'faraday'
 require 'faraday_middleware'
 require 'logger'
+require 'retriable'
 require 'virtus'
 
 path = File.join(File.expand_path(File.dirname(__FILE__)), 'survey_gizmo')
